Provide a zmqchan.ChannelPair class
This patch provides an experimental zmqchan.ChannelPair class which wraps a
zmq socket (as provided by https://github.com/pebbe/zmq4 ) into a TX and
RX channel.

Currently ZMQ sockets are not threadsafe. These are difficult to use in
combination with golang channels as you can poll on a set of sockets, or
select on a set of channels, but not both. This creates problems if you
want to use conventional go techniques, e.g. using a chan bool for ending
goroutines.

This experimental library provides a means of wrapping a ZMQ socket into
a ChannelPair, which provides an Rx and Tx channel (as well as an error
channel). This is loosely based on the idea of another go binding:
   http://github.com/vaughan0/go-zmq
but works with ZMQ 4.x.

This is currently lightly tested / experimental.

Signed-off-by: Alex Bligh <alex@alex.org.uk>